Completed migration of Binary Tree problems
