+++ docs/binarytree/README.md
@@ -1,13 +1,13 @@
 ## Binary Tree
 
-1.  [In-order Traversal of Binary Tree](Medium/InorderTraversalOfBinaryTree)
-2.  [Pre-order Traversal of Binary Tree](Easy/PreorderTraversalOfBinaryTree)
-3.  [Post-order Traversal of Binary Tree](Hard/PostorderTraversalOfBinaryTree)
-4.  [Check if Binary Tree is Balanced](Medium/CheckIfBinaryTreeIsBalanced)
-5.  [Symmetric Binary Tree](Easy/SymmetricBinaryTree)
-6.  [Tweaked Identical Binary Tree](Medium/TweakedIdenticalBinaryTree)
-7.  [Is Binary Search Tree or Not](Medium/IsBinarySearchTreeOrNot)
-8.  [Get Keys in Binary Search Tree in Given Range](Easy/GetKeysInBinarySearchTreeInGivenRange)
-9.  [Binary Tree Path Sum to Target I](Easy/BinaryTreePathSumToTargetI)
-10. [Search in Binary Search Tree](Easy/SearchInBinarySearchTree)
-11. [Insert in Binary Search Tree](Easy/InsertInBinarySearchTree)
+1.  [In-order Traversal of Binary Tree](InorderTraversalOfBinaryTree)
+2.  [Pre-order Traversal of Binary Tree](PreorderTraversalOfBinaryTree)
+3.  [Post-order Traversal of Binary Tree](PostorderTraversalOfBinaryTree)
+4.  [Check if Binary Tree is Balanced](CheckIfBinaryTreeIsBalanced)
+5.  [Symmetric Binary Tree](SymmetricBinaryTree)
+6.  [Tweaked Identical Binary Trees](TweakedIdenticalBinaryTrees)
+7.  [Is Binary Search Tree or Not](IsBinarySearchTreeOrNot)
+8.  [Get Keys in Binary Search Tree in Given Range](GetKeysInBinarySearchTreeInGivenRange)
+9.  [Binary Tree Path Sum to Target I](BinaryTreePathSumToTargetI)
+10. [Search in Binary Search Tree](SearchInBinarySearchTree)
+11. [Insert in Binary Search Tree](InsertInBinarySearchTree)
